+++ apps/vs-code-designer/src/app/utils/codeless/startDesignTimeApi.ts
@@ -86,7 +86,7 @@ export async function startDesignTimeApi(projectPath: string): Promise<void> {
       );
 
       const designTimeDirectory: Uri | undefined = await getOrCreateDesignTimeDirectory(designTimeDirectoryName, projectPath);
-      settingsFileContent.Values[ProjectDirectoryPath] = path.join(designTimeDirectory.fsPath);
+      settingsFileContent.Values[ProjectDirectoryPath] = path.join(projectPath);
 
       if (designTimeDirectory) {
         await createJsonFile(designTimeDirectory, hostFileName, hostFileContent);
@@ -114,6 +114,10 @@ export async function startDesignTimeApi(projectPath: string): Promise<void> {
 
 export async function getOrCreateDesignTimeDirectory(designTimeDirectory: string, projectRoot: string): Promise<Uri | undefined> {
   const directory: string = designTimeDirectory + path.sep;
+  if (projectRoot.includes(designTimeDirectoryName)) {
+    return Uri.file(projectRoot);
+  }
+
   const designTimeDirectoryUri: Uri = Uri.file(path.join(projectRoot, directory));
   if (!fs.existsSync(designTimeDirectoryUri.fsPath)) {
     await workspace.fs.createDirectory(designTimeDirectoryUri);
@@ -190,12 +194,13 @@ export function startDesignTimeProcess(
 }
 
 export function stopDesignTimeApi(): void {
+  ext.outputChannel.appendLog('Stopping Design Time Api');
   if (ext.designChildProcess === null || ext.designChildProcess === undefined) {
     return;
   }
 
   if (os.platform() === Platform.windows) {
-    cp.exec('taskkill /pid ' + `${ext.designChildProcess.pid}` + ' /T /F');
+    cp.exec('taskkill /pid ' + `${ext.designChildProcess.pid}` + ' /t /f');
   } else {
     ext.designChildProcess.kill();
   }