+++ libs/designer-ui/src/lib/editor/base/EditorWrapper.tsx
@@ -1,20 +1,25 @@
 import { BaseEditor, type BaseEditorProps } from '../..';
 import { defaultInitialConfig, defaultNodes, htmlNodes } from './utils/initialConfig';
+import type { SegmentParserOptions } from './utils/parsesegments';
 import { parseHtmlSegments, parseSegments } from './utils/parsesegments';
 import type { InitialConfigType } from '@lexical/react/LexicalComposer';
 import { LexicalComposer } from '@lexical/react/LexicalComposer';
 
 export const EditorWrapper = ({ ...props }: BaseEditorProps) => {
   const { initialValue, basePlugins = {}, readonly } = props;
-  const { isHtmlEditor, tokens } = basePlugins;
+  const { htmlEditor, tokens = true } = basePlugins;
   const initialConfig: InitialConfigType = {
     ...defaultInitialConfig,
     editable: !readonly,
-    nodes: isHtmlEditor ? htmlNodes : defaultNodes,
+    nodes: htmlEditor ? htmlNodes : defaultNodes,
     editorState:
       initialValue &&
       (() => {
-        isHtmlEditor ? parseHtmlSegments(initialValue, tokens, readonly) : parseSegments(initialValue, tokens, readonly);
+        const options: SegmentParserOptions = {
+          readonly,
+          tokensEnabled: tokens,
+        };
+        htmlEditor === 'rich-html' ? parseHtmlSegments(initialValue, options) : parseSegments(initialValue, options);
       }),
   };
   return (