+++ libs/designer-ui/src/lib/editor/base/index.tsx
@@ -3,6 +3,7 @@ import { DESELECT_NODE } from '../../token/inputToken';
 import type { TokenPickerMode } from '../../tokenpicker';
 import { useId } from '../../useId';
 import type { ValueSegment } from '../models/parameter';
+import { ArrowNavigation } from './plugins/ArrowNavigation';
 import { AutoFocus } from './plugins/AutoFocus';
 import AutoLink from './plugins/AutoLink';
 import ClearEditor from './plugins/ClearEditor';
@@ -57,7 +58,7 @@ export interface BaseEditorProps {
   className?: string;
   readonly?: boolean;
   placeholder?: string;
-  basePlugins?: basePlugins;
+  basePlugins?: BasePlugins;
   initialValue: ValueSegment[];
   children?: React.ReactNode;
   labelId?: string;
@@ -71,16 +72,17 @@ export interface BaseEditorProps {
   onBlur?: () => void;
   onFocus?: () => void;
   getTokenPicker: GetTokenPickerHandler;
+  setIsValuePlaintext?: (isValuePlaintext: boolean) => void;
 }
 
-export interface basePlugins {
+export interface BasePlugins {
   autoFocus?: boolean;
   autoLink?: boolean;
   clearEditor?: boolean;
   history?: boolean;
   tokens?: boolean;
   treeView?: boolean;
-  isHtmlEditor?: boolean;
+  htmlEditor?: 'rich-html' | 'raw-html' | false;
   tabbable?: boolean;
   singleValueSegment?: boolean;
 }
@@ -100,6 +102,7 @@ export const BaseEditor = ({
   onFocus,
   onBlur,
   getTokenPicker,
+  setIsValuePlaintext,
 }: BaseEditorProps) => {
   const [editor] = useLexicalComposerContext();
   const editorId = useId('msla-tokenpicker-callout-location');
@@ -128,9 +131,9 @@ export const BaseEditor = ({
     autoLink,
     clearEditor,
     history = true,
-    tokens,
+    tokens = true,
     treeView,
-    isHtmlEditor = false,
+    htmlEditor = false,
     tabbable,
     singleValueSegment = false,
   } = basePlugins;
@@ -165,7 +168,7 @@ export const BaseEditor = ({
   };
 
   const id = useId('msla-described-by-message');
-  const TextPlugin = isHtmlEditor ? RichTextPlugin : PlainTextPlugin;
+  const TextPlugin = htmlEditor === 'rich-html' ? RichTextPlugin : PlainTextPlugin;
 
   return (
     <>
@@ -176,7 +179,7 @@ export const BaseEditor = ({
         data-automation-id={dataAutomationId}
         title={placeholder}
       >
-        {isHtmlEditor ? <Toolbar readonly={readonly} /> : null}
+        {htmlEditor ? <Toolbar isRawText={htmlEditor === 'raw-html'} readonly={readonly} setIsRawText={setIsValuePlaintext} /> : null}
         <TextPlugin
           contentEditable={
             <ContentEditable className={css('editor-input', readonly && 'readonly')} ariaLabelledBy={labelId} ariaDescribedBy={id} />
@@ -197,24 +200,27 @@ export const BaseEditor = ({
         {autoLink ? <AutoLink /> : null}
         {clearEditor ? <ClearEditor showButton={false} /> : null}
         {singleValueSegment ? <SingleValueSegment /> : null}
-        {tokens ? (
-          <TokenTypeAheadPlugin
-            openTokenPicker={openTokenPicker}
-            isEditorFocused={isEditorFocused}
-            hideTokenPickerOptions={tokenPickerButtonProps?.hideButtonOptions}
-          />
-        ) : null}
         <FocusChangePlugin onFocus={handleFocus} onBlur={handleBlur} onClick={handleClick} />
         <ReadOnly readonly={readonly} />
         {tabbable ? null : <IgnoreTab />}
-        {tokens ? <InsertTokenNode /> : null}
-        {tokens ? <DeleteTokenNode /> : null}
-        {tokens ? <OpenTokenPicker openTokenPicker={openTokenPicker} /> : null}
-        {tokens ? <CloseTokenPicker closeTokenPicker={() => setIsTokenPickerOpened(false)} /> : null}
-        {tokens && !isHtmlEditor ? (
+        {htmlEditor === 'rich-html' ? null : <ArrowNavigation />}
+        {tokens ? (
+          <>
+            <InsertTokenNode />
+            <DeleteTokenNode />
+            <OpenTokenPicker openTokenPicker={openTokenPicker} />
+            <CloseTokenPicker closeTokenPicker={() => setIsTokenPickerOpened(false)} />
+            <TokenTypeAheadPlugin
+              openTokenPicker={openTokenPicker}
+              isEditorFocused={isEditorFocused}
+              hideTokenPickerOptions={tokenPickerButtonProps?.hideButtonOptions}
+            />
+          </>
+        ) : null}
+        {tokens && !htmlEditor ? (
           <PastePlugin segmentMapping={tokenMapping} loadParameterValueFromString={loadParameterValueFromString} />
         ) : null}
-        {isHtmlEditor && floatingAnchorElem ? <FloatingLinkEditorPlugin anchorElem={floatingAnchorElem} /> : null}
+        {htmlEditor && floatingAnchorElem ? <FloatingLinkEditorPlugin anchorElem={floatingAnchorElem} /> : null}
         {children}
         {tokens && isTokenPickerOpened ? getTokenPicker(editorId, labelId ?? '', tokenPickerMode, valueType, setIsTokenPickerOpened) : null}
       </div>