+++ libs/designer-ui/src/lib/picker/plugins/UpdateEditorFromFilePicker.tsx
@@ -8,14 +8,14 @@ interface ChangeProps {
   pickerDisplayValue: ValueSegment[];
 }
 
-export const UpdateEditorFromPicker = ({ pickerDisplayValue }: ChangeProps) => {
+export const UpdateEditorFromFilePicker = ({ pickerDisplayValue }: ChangeProps) => {
   const [editor] = useLexicalComposerContext();
 
   useUpdateEffect(() => {
     if (pickerDisplayValue.length > 0) {
       editor.dispatchCommand(CLEAR_EDITOR_COMMAND, undefined);
       editor.update(() => {
-        parseSegments(pickerDisplayValue, /*tokensEnabled*/ true, /*readonly*/ false);
+        parseSegments(pickerDisplayValue, { tokensEnabled: true, readonly: false });
       });
     }
   }, [editor, pickerDisplayValue]);