+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/nodeDetailsPanel.tsx
@@ -13,40 +13,31 @@ import { deleteGraphNode, deleteOperation } from '../../../core/actions/bjsworkf
 import type { WorkflowNode } from '../../../core/parsers/models/workflowNode';
 import { useReadOnly } from '../../../core/state/designerOptions/designerOptionsSelectors';
 import { ErrorLevel } from '../../../core/state/operation/operationMetadataSlice';
-import { useOperationErrorInfo, useParameterValidationErrors } from '../../../core/state/operation/operationSelector';
-import {
-  useIsPanelCollapsed,
-  useCurrentPanelMode,
-  useRegisteredPanelTabs,
-  useSelectedPanelTabName,
-} from '../../../core/state/panel/panelSelectors';
-import { setTabError, expandPanel, selectPanelTab, updatePanelLocation } from '../../../core/state/panel/panelSlice';
+import { useOperationErrorInfo } from '../../../core/state/operation/operationSelector';
+import { useIsPanelCollapsed, useSelectedPanelTabId } from '../../../core/state/panel/panelSelectors';
+import { expandPanel, updatePanelLocation, selectPanelTab } from '../../../core/state/panel/panelSlice';
 import { useIconUri, useOperationQuery } from '../../../core/state/selectors/actionMetadataSelector';
 import { useWorkflowNode, useNodeDescription } from '../../../core/state/workflow/workflowSelectors';
 import { deleteSwitchCase, setNodeDescription, replaceId } from '../../../core/state/workflow/workflowSlice';
 import { isRootNodeInGraph, isOperationNameValid } from '../../../core/utils/graph';
-import type { CommonPanelProps, MenuItemOption, PageActionTelemetryData } from '@microsoft/designer-ui';
-import {
-  DeleteNodeModal,
-  MenuItemType,
-  PanelContainer,
-  PanelHeaderControlType,
-  PanelLocation,
-  PanelScope,
-  PanelSize,
-} from '@microsoft/designer-ui';
+import { CommentMenuItem } from '../../menuItems/commentMenuItem';
+import { DeleteMenuItem } from '../../menuItems/deleteMenuItem';
+import { usePanelTabs } from './usePanelTabs';
+import type { CommonPanelProps, PageActionTelemetryData } from '@microsoft/designer-ui';
+import { DeleteNodeModal, PanelContainer, PanelLocation, PanelScope, PanelSize } from '@microsoft/designer-ui';
 import { isNullOrUndefined, isScopeOperation, isSubGraphNode } from '@microsoft/utils-logic-apps';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 
 export const NodeDetailsPanel = (props: CommonPanelProps): JSX.Element => {
   const { panelLocation } = props;
 
-  const intl = useIntl();
   const dispatch = useDispatch<AppDispatch>();
 
   const readOnly = useReadOnly();
+
+  const panelTabs = usePanelTabs();
+  const selectedTab = useSelectedPanelTabId();
 
   const collapsed = useIsPanelCollapsed();
   const selectedNode = useSelectedNodeId();
@@ -56,17 +47,11 @@ export const NodeDetailsPanel = (props: CommonPanelProps): JSX.Element => {
     idReplacements: state.workflow.idReplacements,
   }));
   const selectedNodeDisplayName = useNodeDisplayName(selectedNode);
-  const currentPanelMode = useCurrentPanelMode();
 
   const graphNode = useWorkflowNode(selectedNode) as WorkflowNode;
 
   const [width, setWidth] = useState<PanelSize>(PanelSize.Auto);
 
-  const registeredTabs = useRegisteredPanelTabs();
-  const selectedPanelTab = useSelectedPanelTabName();
-  const setSelectedTab = (tabName: string | undefined) => {
-    dispatch(selectPanelTab(tabName));
-  };
   const inputs = useSelector((state: RootState) => state.operations.inputParameters[selectedNode]);
   const comment = useNodeDescription(selectedNode);
   const iconUri = useIconUri(selectedNode);
@@ -74,12 +59,6 @@ export const NodeDetailsPanel = (props: CommonPanelProps): JSX.Element => {
   const operationInfo = useOperationInfo(selectedNode);
   let showCommentBox = !isNullOrUndefined(comment);
   const errorInfo = useOperationErrorInfo(selectedNode);
-
-  const parameterValidationErrors = useParameterValidationErrors(selectedNode);
-  useEffect(() => {
-    const hasErrors = parameterValidationErrors?.length > 0 || errorInfo?.level === ErrorLevel.Connection;
-    dispatch(setTabError({ tabName: 'parameters', hasErrors, nodeId: selectedNode }));
-  }, [dispatch, errorInfo?.level, parameterValidationErrors?.length, selectedNode]);
 
   useEffect(() => {
     collapsed ? setWidth(PanelSize.Auto) : setWidth(PanelSize.Medium);
@@ -112,76 +91,15 @@ export const NodeDetailsPanel = (props: CommonPanelProps): JSX.Element => {
     setShowDeleteModal(false);
   };
 
-  const getPanelHeaderControlType = (): boolean => {
-    // TODO: 13067650
-    return currentPanelMode === 'Discovery';
-  };
-
-  const getPanelHeaderMenu = (): MenuItemOption[] => {
-    const menuOptions: MenuItemOption[] = [];
-    // TODO: 13067650 Conditionals to decide when to show these menu options
-    getCommentMenuItem(menuOptions);
-    getDeleteMenuItem(menuOptions);
-    return menuOptions;
-  };
-
-  const getCommentMenuItem = (options: MenuItemOption[]): MenuItemOption[] => {
-    const commentDescription = intl.formatMessage({
-      defaultMessage: 'Note',
-      description: 'Note text',
-    });
-    const disabledCommentAction = intl.formatMessage({
-      defaultMessage: "You can add notes only when you edit a step's inputs.",
-      description: 'Text to tell users why notes are disabled',
-    });
-    const commentAdd = intl.formatMessage({
-      defaultMessage: 'Add a note',
-      description: 'Text to tell users to click to add comments',
-    });
-    const commentDelete = intl.formatMessage({
-      defaultMessage: 'Delete note',
-      description: 'Text to tell users to click to delete comments',
-    });
-
-    options.push({
-      disabled: readOnly,
-      type: MenuItemType.Advanced,
-      disabledReason: disabledCommentAction,
-      iconName: 'Comment',
-      key: commentDescription,
-      title: showCommentBox ? commentDelete : commentAdd,
-      onClick: handleCommentMenuClick,
-    });
-    return options;
-  };
-
-  const getDeleteMenuItem = (options: MenuItemOption[]): MenuItemOption[] => {
-    const deleteDescription = intl.formatMessage({
-      defaultMessage: 'Delete',
-      description: 'Delete text',
-    });
-    // TODO: 13067650 Disabled reason/description tobe implemented when panel actions gets built
-    const disabledDeleteAction = intl.formatMessage({
-      defaultMessage: 'This operation has already been deleted.',
-      description: 'Text to tell users why delete is disabled',
-    });
-
-    options.push({
-      key: deleteDescription,
-      disabled: readOnly,
-      disabledReason: disabledDeleteAction,
-      iconName: 'Delete',
-      title: deleteDescription,
-      type: MenuItemType.Advanced,
-      onClick: handleDeleteClick,
-    });
-    return options;
-  };
-
   const handleCommentMenuClick = (_: React.MouseEvent<HTMLElement>): void => {
     showCommentBox = !showCommentBox;
     dispatch(setNodeDescription({ nodeId: selectedNode, ...(showCommentBox && { description: '' }) }));
   };
+
+  const headerMenuItems = [
+    <CommentMenuItem key={'comment'} onClick={handleCommentMenuClick} hasComment={showCommentBox} />,
+    <DeleteMenuItem key={'delete'} onClick={handleDeleteClick} />,
+  ];
 
   const onTitleChange = (newId: string): { valid: boolean; oldValue?: string } => {
     const isValid = isOperationNameValid(selectedNode, newId, isTriggerNode, nodesMetadata, idReplacements);
@@ -228,15 +146,13 @@ export const NodeDetailsPanel = (props: CommonPanelProps): JSX.Element => {
         errorMessage={errorInfo?.message}
         isLoading={isLoading}
         panelScope={PanelScope.CardLevel}
-        panelHeaderControlType={getPanelHeaderControlType() ? PanelHeaderControlType.DISMISS_BUTTON : PanelHeaderControlType.MENU}
-        panelHeaderMenu={getPanelHeaderMenu()}
-        selectedTab={selectedPanelTab}
+        headerMenuItems={headerMenuItems}
         showCommentBox={showCommentBox}
-        tabs={registeredTabs}
+        tabs={panelTabs}
+        selectedTab={selectedTab}
+        selectTab={(tabId: string) => dispatch(selectPanelTab(tabId))}
         nodeId={selectedNode}
-        onDismissButtonClicked={handleDelete}
         readOnlyMode={readOnly}
-        setSelectedTab={setSelectedTab}
         toggleCollapse={() => {
           // Only run validation when collapsing the panel
           if (!collapsed) {