+++ libs/designer/src/lib/ui/settings/sections/runafterconfiguration/runafteractionstatuses.tsx
@@ -1,76 +1,70 @@
 import type { onChangeHandler } from './runafteractiondetails';
-import { Checkbox } from '@fluentui/react/lib/Checkbox';
+import { Text, Checkbox } from '@fluentui/react-components';
+import { Failed, Skipped, Succeeded, TimedOut } from '@microsoft/designer-ui';
 import { RUN_AFTER_STATUS } from '@microsoft/utils-logic-apps';
 import { useIntl } from 'react-intl';
 
 export interface RunAfterActionStatusesProps {
   isReadOnly: boolean;
   statuses: string[];
-  onRenderLabel(status: string, label: string): JSX.Element;
   onStatusChange?: onChangeHandler;
 }
 
-export function RunAfterActionStatuses({ isReadOnly, statuses, onStatusChange, onRenderLabel }: RunAfterActionStatusesProps): JSX.Element {
+export function RunAfterActionStatuses({ isReadOnly, statuses, onStatusChange }: RunAfterActionStatusesProps): JSX.Element {
   const intl = useIntl();
   const normalizedStatuses = statuses.map((status) => status.toUpperCase());
 
-  const Resources = {
-    RUN_AFTER_SUCCEEDED_STATUS: intl.formatMessage({
-      defaultMessage: 'Is successful',
-      description: 'Successful run',
-    }),
-    RUN_AFTER_TIMEDOUT_STATUS: intl.formatMessage({
-      defaultMessage: 'Has timed out',
-      description: 'Timed out run',
-    }),
-    RUN_AFTER_SKIPPED_STATUS: intl.formatMessage({
-      defaultMessage: 'Is skipped',
-      description: 'Skipped run',
-    }),
-    RUN_AFTER_FAILED_STATUS: intl.formatMessage({
-      defaultMessage: 'Has failed',
-      description: 'Failed run',
-    }),
-  };
+  const data = [
+    {
+      id: RUN_AFTER_STATUS.SUCCEEDED,
+      label: intl.formatMessage({
+        defaultMessage: 'Is successful',
+        description: 'Successful run',
+      }),
+      icon: <Succeeded />,
+    },
+    {
+      id: RUN_AFTER_STATUS.TIMEDOUT,
+      label: intl.formatMessage({
+        defaultMessage: 'Has timed out',
+        description: 'Timed out run',
+      }),
+      icon: <TimedOut />,
+    },
+    {
+      id: RUN_AFTER_STATUS.SKIPPED,
+      label: intl.formatMessage({
+        defaultMessage: 'Is skipped',
+        description: 'Skipped run',
+      }),
+      icon: <Skipped />,
+    },
+    {
+      id: RUN_AFTER_STATUS.FAILED,
+      label: intl.formatMessage({
+        defaultMessage: 'Has failed',
+        description: 'Failed run',
+      }),
+      icon: <Failed />,
+    },
+  ];
 
   return (
     <div className="msla-run-after-statuses">
-      <div className="msla-run-after-status-checkbox">
+      {data.map((status) => (
         <Checkbox
-          checked={normalizedStatuses.includes(RUN_AFTER_STATUS.SUCCEEDED)}
+          key={status.id}
+          checked={normalizedStatuses.includes(status.id)}
           disabled={isReadOnly}
-          label={Resources.RUN_AFTER_SUCCEEDED_STATUS}
-          onChange={(_, checked) => onStatusChange?.(RUN_AFTER_STATUS.SUCCEEDED, checked)}
-          onRenderLabel={() => onRenderLabel(RUN_AFTER_STATUS.SUCCEEDED, Resources.RUN_AFTER_SUCCEEDED_STATUS)}
+          label={
+            <div className="status-label">
+              {status.icon}
+              <Text>{status.label}</Text>
+            </div>
+          }
+          onChange={(_, data) => onStatusChange?.(status.id, !!data.checked)}
         />
-      </div>
-      <div className="msla-run-after-status-checkbox">
-        <Checkbox
-          checked={normalizedStatuses.includes(RUN_AFTER_STATUS.TIMEDOUT)}
-          disabled={isReadOnly}
-          label={Resources.RUN_AFTER_TIMEDOUT_STATUS}
-          onChange={(_, checked) => onStatusChange?.(RUN_AFTER_STATUS.TIMEDOUT, checked)}
-          onRenderLabel={() => onRenderLabel(RUN_AFTER_STATUS.TIMEDOUT, Resources.RUN_AFTER_TIMEDOUT_STATUS)}
-        />
-      </div>
-      <div className="msla-run-after-status-checkbox">
-        <Checkbox
-          checked={normalizedStatuses.includes(RUN_AFTER_STATUS.SKIPPED)}
-          disabled={isReadOnly}
-          label={Resources.RUN_AFTER_SKIPPED_STATUS}
-          onChange={(_, checked) => onStatusChange?.(RUN_AFTER_STATUS.SKIPPED, checked)}
-          onRenderLabel={() => onRenderLabel(RUN_AFTER_STATUS.SKIPPED, Resources.RUN_AFTER_SKIPPED_STATUS)}
-        />
-      </div>
-      <div className="msla-run-after-status-checkbox">
-        <Checkbox
-          checked={normalizedStatuses.includes(RUN_AFTER_STATUS.FAILED)}
-          disabled={isReadOnly}
-          label={Resources.RUN_AFTER_FAILED_STATUS}
-          onChange={(_, checked) => onStatusChange?.(RUN_AFTER_STATUS.FAILED, checked)}
-          onRenderLabel={() => onRenderLabel(RUN_AFTER_STATUS.FAILED, Resources.RUN_AFTER_FAILED_STATUS)}
-        />
-      </div>
+      ))}
     </div>
   );
 }