+++ CHANGELOG.md
@@ -1,33 +1,77 @@
 # Logic Apps Designer
-## [2.93.0](https://github.com/Azure/LogicAppsUX/compare/v2.92.0...v2.93.0) (2023-12-14)
+## [2.99.0](https://github.com/Azure/LogicAppsUX/compare/v2.98.0...v2.99.0) (2024-01-02)
 
 
 ### Bug Fixes
 
-* **designer-ui:** Fix pill tooltip UX and corner rounding ([#3843](https://github.com/Azure/LogicAppsUX/issues/3843)) ([c06f9a5](https://github.com/Azure/LogicAppsUX/commit/c06f9a5ec5845de430acf84123fe78ceb3cfde45))
-* **Designer:** Connections panel refactor ([#3815](https://github.com/Azure/LogicAppsUX/issues/3815)) ([92d28ee](https://github.com/Azure/LogicAppsUX/commit/92d28ee64116b765c8c483b40fdf02b9ae6a81e2))
+* **vscode:** Add validation for design-time folder ([#3910](https://github.com/Azure/LogicAppsUX/issues/3910)) ([0c48a29](https://github.com/Azure/LogicAppsUX/commit/0c48a297e39403351dc6cdc3aefe3e1735cff13c))
 
-## [2.92.0](https://github.com/Azure/LogicAppsUX/compare/v2.91.0...v2.92.0) (2023-12-14)
+## [2.98.0](https://github.com/Azure/LogicAppsUX/compare/v2.97.0...v2.98.0) (2023-12-30)
+
+## [2.97.0](https://github.com/Azure/LogicAppsUX/compare/v2.96.0...v2.97.0) (2023-12-28)
+
+## [2.96.0](https://github.com/Azure/LogicAppsUX/compare/v2.95.0...v2.96.0) (2023-12-21)
 
 
 ### Features
 
-* **vscode:** Remove designer refresh and add window message ([#3816](https://github.com/Azure/LogicAppsUX/issues/3816)) ([8633353](https://github.com/Azure/LogicAppsUX/commit/8633353c6ef66510f6379caf7d26c8d09067e522))
+* **designer:** add support prevent casting in complex array editors ([#3873](https://github.com/Azure/LogicAppsUX/issues/3873)) ([5d96b77](https://github.com/Azure/LogicAppsUX/commit/5d96b77055776a3eb4a7067559fa03626d16a8a5))
+* **designer:** enable support prevent casting in complex array editors ([#3886](https://github.com/Azure/LogicAppsUX/issues/3886)) ([ff0af65](https://github.com/Azure/LogicAppsUX/commit/ff0af65796ae24d562674caf0b913206cc422313))
 
 
 ### Bug Fixes
 
-* **copilot:** Raising Telemetry Level of Copilot ([#3832](https://github.com/Azure/LogicAppsUX/issues/3832)) ([e787a25](https://github.com/Azure/LogicAppsUX/commit/e787a25b33fab39eb9afef95881c86c9bc53cab2))
-* **Data Mapper:** Disabled allowing undo add schema ([#3837](https://github.com/Azure/LogicAppsUX/issues/3837)) ([f852e2a](https://github.com/Azure/LogicAppsUX/commit/f852e2a9d6d6e02d9c619b0d3b474328ff9e3414))
-* **Data Mapper:** Display Error Message for Failing Schema Load ([#3839](https://github.com/Azure/LogicAppsUX/issues/3839)) ([4146cdc](https://github.com/Azure/LogicAppsUX/commit/4146cdc470e1f24bcf57e73866f72bb9ab724580))
-* **designer:** Add connector type info to about panel and refactor how displayRuntimeInfo is passed in ([#3812](https://github.com/Azure/LogicAppsUX/issues/3812)) ([c2b2ef3](https://github.com/Azure/LogicAppsUX/commit/c2b2ef3e7e348fd3da843044201fb44e0f8178e8))
-* **designer:** Add default text color for token picker pivot ([#3838](https://github.com/Azure/LogicAppsUX/issues/3838)) ([11ff584](https://github.com/Azure/LogicAppsUX/commit/11ff58428c0c06448ca43bbfe3b3ae6475edbc1e))
-* **designer:** Fix toggle for collapsed and expanded dictionary view … ([#3811](https://github.com/Azure/LogicAppsUX/issues/3811)) ([b5690d5](https://github.com/Azure/LogicAppsUX/commit/b5690d52b157b283ea8f8ef24197d345b46bb72b))
-* **designer:** Fixes an issue where we weren't properly creating new lines in the editor ([#3823](https://github.com/Azure/LogicAppsUX/issues/3823)) ([67ccddf](https://github.com/Azure/LogicAppsUX/commit/67ccddf3b3d0f48b6bcee6daa302ee2f7f833ef5))
-* **designer:** issue where designer would constantly rerender if anything about hte window changed ([#3814](https://github.com/Azure/LogicAppsUX/issues/3814)) ([de11487](https://github.com/Azure/LogicAppsUX/commit/de114872891188cd45c4772a074b2a90c5da2e45))
-* **vscode:** Add troubleshoot dependencies message [#3687](https://github.com/Azure/LogicAppsUX/issues/3687) ([#3833](https://github.com/Azure/LogicAppsUX/issues/3833)) ([992eb3f](https://github.com/Azure/LogicAppsUX/commit/992eb3f9356977f1e906bbbde1d279e167bc0197))
-* **vscode:** Add workflow app kind to local.settings.json ([#3841](https://github.com/Azure/LogicAppsUX/issues/3841)) ([852b535](https://github.com/Azure/LogicAppsUX/commit/852b535e030730d1aadeb2f150318aa78ab2e24c))
-* **vscode:** Check list of files in dotnet binary installation ([#3813](https://github.com/Azure/LogicAppsUX/issues/3813)) ([319e660](https://github.com/Azure/LogicAppsUX/commit/319e6604a9b7ac250df2e75dfd570d8ddc916f9b))
+* **designer:** Add description for Condition action ([#3865](https://github.com/Azure/LogicAppsUX/issues/3865)) ([aec1286](https://github.com/Azure/LogicAppsUX/commit/aec12861bcb485d293b75b3468b9be0deda39605))
+* **designer:** Changes not serializing when token deleted without editor being focused ([#3882](https://github.com/Azure/LogicAppsUX/issues/3882)) ([9f73a3c](https://github.com/Azure/LogicAppsUX/commit/9f73a3c331571ce897d847f7e2ffda49de8f38ec))
+* **Designer:** Converted many fluent8 components to fluent9 ([#3872](https://github.com/Azure/LogicAppsUX/issues/3872)) ([9d06a1e](https://github.com/Azure/LogicAppsUX/commit/9d06a1edff0445053ee09b6d7025871b5528b336))
+* **designer:** Fix regressions moving token parsing from  $[ ]$ to @{} ([#3881](https://github.com/Azure/LogicAppsUX/issues/3881)) ([e005c8d](https://github.com/Azure/LogicAppsUX/commit/e005c8d5116dcef3237a1948888d0dc200c0f98b))
+* **designer:** Fixed issue where lexicalPlainText did not have proper keyboard navigation for token nodes ([#3854](https://github.com/Azure/LogicAppsUX/issues/3854)) ([6d91c38](https://github.com/Azure/LogicAppsUX/commit/6d91c38450e4da7cc6f2482a8633dee5788035e9))
+* **designer:** Revert change "tokens with 'body/value' should use triggerBody rather than triggerOutput" ([#3875](https://github.com/Azure/LogicAppsUX/issues/3875)) ([a09c39a](https://github.com/Azure/LogicAppsUX/commit/a09c39a7687fba0ee57b6d734299542fcce04cf8)), closes [#3782](https://github.com/Azure/LogicAppsUX/issues/3782)
+* **designer:** Update condition on when to render monitoring view ([#3892](https://github.com/Azure/LogicAppsUX/issues/3892)) ([29cd1f8](https://github.com/Azure/LogicAppsUX/commit/29cd1f8f43b3611e5e2054c1381b52f5c3de5138))
+* **localization:** Adding Indonesian as a language to intl provider ([#3868](https://github.com/Azure/LogicAppsUX/issues/3868)) ([40bc6f8](https://github.com/Azure/LogicAppsUX/commit/40bc6f894e38e955ff19a8fd942e36a84d5fe0bd))
+* **vscode:** Add validate input for function name in workspace creation ([#3874](https://github.com/Azure/LogicAppsUX/issues/3874)) ([7abb3bf](https://github.com/Azure/LogicAppsUX/commit/7abb3bf663b79202cda4991252bf9e48137aa4d2))
+* **vscode:** Dependency timeout ([#3846](https://github.com/Azure/LogicAppsUX/issues/3846)) ([e87a557](https://github.com/Azure/LogicAppsUX/commit/e87a5574d9d7e77a42596a9615b37992e5fd9aba))
+* **vscode:** Fix project path in workflow-designtime folder ([#3867](https://github.com/Azure/LogicAppsUX/issues/3867)) ([712c4ff](https://github.com/Azure/LogicAppsUX/commit/712c4ff609ac8ea39b1231712121060e8a2a71d1))
+
+## [2.95.0](https://github.com/Azure/LogicAppsUX/compare/v2.94.0...v2.95.0) (2023-12-15)
+
+### Features
+
+- **designer-ui:** Support raw code view in HTML editor ([#3849](https://github.com/Azure/LogicAppsUX/issues/3849)) ([13a76cc](https://github.com/Azure/LogicAppsUX/commit/13a76cc1be46c98e91526e887834a4e90d9b34ed))
+
+## [2.94.0](https://github.com/Azure/LogicAppsUX/compare/v2.93.0...v2.94.0) (2023-12-15)
+
+### Bug Fixes
+
+- **Data Mapper:** fixing edifact nested loop mapping bug ([#3824](https://github.com/Azure/LogicAppsUX/issues/3824)) ([575f61d](https://github.com/Azure/LogicAppsUX/commit/575f61d22e9a413ddb30baea3bc582c6592ba548))
+- **Designer:** Git-Removed the .vscode/settings.json file ([#3847](https://github.com/Azure/LogicAppsUX/issues/3847)) ([c769a5e](https://github.com/Azure/LogicAppsUX/commit/c769a5e83ce24c456f679f9ed42262a82ba02e86))
+
+## [2.93.0](https://github.com/Azure/LogicAppsUX/compare/v2.92.0...v2.93.0) (2023-12-14)
+
+### Bug Fixes
+
+- **designer-ui:** Fix pill tooltip UX and corner rounding ([#3843](https://github.com/Azure/LogicAppsUX/issues/3843)) ([c06f9a5](https://github.com/Azure/LogicAppsUX/commit/c06f9a5ec5845de430acf84123fe78ceb3cfde45))
+- **Designer:** Connections panel refactor ([#3815](https://github.com/Azure/LogicAppsUX/issues/3815)) ([92d28ee](https://github.com/Azure/LogicAppsUX/commit/92d28ee64116b765c8c483b40fdf02b9ae6a81e2))
+
+## [2.92.0](https://github.com/Azure/LogicAppsUX/compare/v2.91.0...v2.92.0) (2023-12-14)
+
+### Features
+
+- **vscode:** Remove designer refresh and add window message ([#3816](https://github.com/Azure/LogicAppsUX/issues/3816)) ([8633353](https://github.com/Azure/LogicAppsUX/commit/8633353c6ef66510f6379caf7d26c8d09067e522))
+
+### Bug Fixes
+
+- **copilot:** Raising Telemetry Level of Copilot ([#3832](https://github.com/Azure/LogicAppsUX/issues/3832)) ([e787a25](https://github.com/Azure/LogicAppsUX/commit/e787a25b33fab39eb9afef95881c86c9bc53cab2))
+- **Data Mapper:** Disabled allowing undo add schema ([#3837](https://github.com/Azure/LogicAppsUX/issues/3837)) ([f852e2a](https://github.com/Azure/LogicAppsUX/commit/f852e2a9d6d6e02d9c619b0d3b474328ff9e3414))
+- **Data Mapper:** Display Error Message for Failing Schema Load ([#3839](https://github.com/Azure/LogicAppsUX/issues/3839)) ([4146cdc](https://github.com/Azure/LogicAppsUX/commit/4146cdc470e1f24bcf57e73866f72bb9ab724580))
+- **designer:** Add connector type info to about panel and refactor how displayRuntimeInfo is passed in ([#3812](https://github.com/Azure/LogicAppsUX/issues/3812)) ([c2b2ef3](https://github.com/Azure/LogicAppsUX/commit/c2b2ef3e7e348fd3da843044201fb44e0f8178e8))
+- **designer:** Add default text color for token picker pivot ([#3838](https://github.com/Azure/LogicAppsUX/issues/3838)) ([11ff584](https://github.com/Azure/LogicAppsUX/commit/11ff58428c0c06448ca43bbfe3b3ae6475edbc1e))
+- **designer:** Fix toggle for collapsed and expanded dictionary view … ([#3811](https://github.com/Azure/LogicAppsUX/issues/3811)) ([b5690d5](https://github.com/Azure/LogicAppsUX/commit/b5690d52b157b283ea8f8ef24197d345b46bb72b))
+- **designer:** Fixes an issue where we weren't properly creating new lines in the editor ([#3823](https://github.com/Azure/LogicAppsUX/issues/3823)) ([67ccddf](https://github.com/Azure/LogicAppsUX/commit/67ccddf3b3d0f48b6bcee6daa302ee2f7f833ef5))
+- **designer:** issue where designer would constantly rerender if anything about hte window changed ([#3814](https://github.com/Azure/LogicAppsUX/issues/3814)) ([de11487](https://github.com/Azure/LogicAppsUX/commit/de114872891188cd45c4772a074b2a90c5da2e45))
+- **vscode:** Add troubleshoot dependencies message [#3687](https://github.com/Azure/LogicAppsUX/issues/3687) ([#3833](https://github.com/Azure/LogicAppsUX/issues/3833)) ([992eb3f](https://github.com/Azure/LogicAppsUX/commit/992eb3f9356977f1e906bbbde1d279e167bc0197))
+- **vscode:** Add workflow app kind to local.settings.json ([#3841](https://github.com/Azure/LogicAppsUX/issues/3841)) ([852b535](https://github.com/Azure/LogicAppsUX/commit/852b535e030730d1aadeb2f150318aa78ab2e24c))
+- **vscode:** Check list of files in dotnet binary installation ([#3813](https://github.com/Azure/LogicAppsUX/issues/3813)) ([319e660](https://github.com/Azure/LogicAppsUX/commit/319e6604a9b7ac250df2e75dfd570d8ddc916f9b))
 
 ## [2.91.0](https://github.com/Azure/LogicAppsUX/compare/v2.90.0...v2.91.0) (2023-12-07)